Cache: Terminate processes when closing executor (#447)
* Cache: Terminate processes when closing executor

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Merge main

* fixes

* refactor cache spawner

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
